handle empty environment properly
Empty environment does not have 2 consequitive zeros because
one of these zeros is from previous ASCIZ env string, which
is not there.
So look also for 0 1, in which case skip only 2 bytes instead of 3.

Note: there are lots of such code instances in api.asm.
But they do not execute for me so this patch does not touch them.
I don't know when they are executed.