check env seg for NULL
Some DOSes set env seg to NULL, either when env is empty or always
for the first process (aka permanent shell).
This patch is trying to gracefully shut down in that case.
Unfortunately, gracefully terminating the initial process is not
always possible, but at least you will see a proper error msg before
hang or a junk output. :)